% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{clipper_BC_2}
\alias{clipper_BC_2}
\title{FDR control method based on clipper (Same as clipper_BC).}
\usage{
clipper_BC_2(contrastScore, FDR)
}
\arguments{
\item{contrastScore}{Contrast Score or fliter statistics;}

\item{FDR}{Scalar or vector, target control level.}
}
\value{
FDR: If is vector, output FDR value of result;
thre: Control thresholds for contrast scores;
q: The converted values of the scores were compared;
discovery: Results of variable selection at the FDR level.
}
\description{
FDR control method based on clipper (Same as clipper_BC).
}
