% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{clipper_p_i_2}
\alias{clipper_p_i_2}
\title{The results of variable selection under different strategies were calculated (clipper_pooling_intersection).}
\usage{
clipper_p_i_2(
  contrast_score_k,
  pooling = FALSE,
  n_p_j,
  statistics = 1,
  FDR = c(0.05, 0.1, 0.2, 0.5, 0.9, 1)
)
}
\arguments{
\item{contrast_score_k}{Test statistics for different sources of multi-source data;}

\item{pooling}{Boolean value, whether variable selection is based on pooling strategy;}

\item{n_p_j}{Same as T_var, known significant variables;}

\item{statistics}{Calculation of Fliter statistics;}

\item{FDR}{Target control level.}
}
\value{
pooling, intersect, and simultaneous: FDR control results under different strategies, c(FDR, fdr_i, power_i);
pooling_p, intersect_p, simultaneous_p: Results of variable selection under different strategies,
}
\description{
The results of variable selection under different strategies were calculated (clipper_pooling_intersection).
}
