% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{clipper_p_i_E}
\alias{clipper_p_i_E}
\title{The results of variable selection under different strategies were calculated (Based on e-value).}
\usage{
clipper_p_i_E(
  E,
  pooling = FALSE,
  n_p_j,
  statistics = 1,
  FDR = c(0.05, 0.1, 0.2, 0.5, 0.9, 1)
)
}
\arguments{
\item{E}{E-value;}

\item{pooling}{Same as 'clipper_p_i_2';}

\item{FDR}{}
}
\description{
The results of variable selection under different strategies were calculated (Based on e-value).
}
