% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{contrast_score_computation}
\alias{contrast_score_computation}
\title{The calculation of test statistics is mainly based on DE's method.}
\usage{
contrast_score_computation(W, result3, class0, test1)
}
\arguments{
\item{W}{Count data;}

\item{result3}{knockoff generation.}

\item{class0}{Response variable, binary variable or others.}

\item{test1}{Selection of test statistics, c("wilcox.test", "ks.test", "mmd", "distance_JS")}
}
\value{
contrast_score: contrast score.
}
\description{
The calculation of test statistics is mainly based on DE's method.
}
