% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_pg_copula}
\alias{data_pg_copula}
\title{Simulated data generation}
\format{
A list of three multi-source datasets with different numbers of variables:
\describe{
\item{n_data}{2 sources}
\item{n_1_all}{Sample Size, c(400, 400)}
\item{n_p_all}{Total number of variables, c(50, 100, 400)}
\item{n_p}{Number of significant variables, c(10, 20, 40)}
}
}
\source{
'simulate_datasets_3'
}
\usage{
data_pg_copula
}
\description{
Simulated data generation
}
\keyword{datasets}
