% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{distance_JS}
\alias{distance_JS}
\title{Calculate JS distance (KL distance)}
\usage{
distance_JS(P, Q)
}
\arguments{
\item{P}{Value of vector;}

\item{Q}{Value of vector.}
}
\description{
Calculate JS distance (KL distance)
}
