% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{ebh}
\alias{ebh}
\title{The e-BH procedure.}
\usage{
ebh(E, alpha)
}
\arguments{
\item{E}{E-value;}

\item{alpha}{Target control level (FDR).}
}
\value{
rej: Results of variable selection.
}
\description{
The e-BH procedure.
}
