% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{ebh_2}
\alias{ebh_2}
\title{The e-BH procedure (different alpha level).}
\usage{
ebh_2(E, alpha)
}
\arguments{
\item{E}{E-value;}

\item{alpha}{A scalar or a vector, target control level (FDR).}
}
\value{
discovery(rej): Results of variable selection.
FDR: alpha_i.
}
\description{
The e-BH procedure (different alpha level).
}
