% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{ekn}
\alias{ekn}
\title{Title}
\usage{
ekn(C, gamma, offset)
}
\arguments{
\item{C}{Vector of knockoff feature importance statistics (test statistics or fliter statistics);}

\item{gamma}{Alpha_kn;}

\item{offset}{Value between 0 and 1.}
}
\value{
E: the list of knockoff e-values.
}
\description{
Title
}
