% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{filter_cw}
\alias{filter_cw}
\title{The values of FDR and Power under other filter statistics (CV).}
\usage{
filter_cw(c_w, filter_statistics = 1, fdr = 0.2, b_1 = NULL)
}
\arguments{
\item{c_w}{Test statistics;}

\item{filter_statistics}{Filter statistics;}

\item{fdr}{Target control level.}

\item{b_1}{True significant variable.}
}
\value{
S: Results of variable selection;
FDRPower: FDR and Power.
}
\description{
The values of FDR and Power under other filter statistics (CV).
}
