% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{filter_cw_B}
\alias{filter_cw_B}
\title{The values of FDR and Power under Aggregating multiple for SKMCD (CV).}
\usage{
filter_cw_B(
  Bstat = 1,
  combine_1 = NULL,
  c_w = NULL,
  e_w = NULL,
  filter_statistics = 1,
  B = 30,
  fdr = 0.2,
  T_var = NULL,
  K = 1
)
}
\arguments{
\item{Bstat}{For aggregation, 1 represents the aggregated test statistic and 2 represents the aggregated fliter statistic;}

\item{combine_1}{When combine_1 aggregation mode is 1, 'simul' means that the statistic is calculated based on the proposed method, and 'fisher' means that the statistic is calculated based on the combination of p values.}

\item{c_w}{Test statistics of B times;}

\item{e_w}{Test statistics (e-value) of B times;}

\item{filter_statistics}{Filter statistics;}

\item{B}{Number of aggregations;}

\item{fdr}{Target control level.}

\item{T_var}{True correlation variable, which represents the set of variables that are truly correlated with the response variable, defaults to NULL;}

\item{K}{Number of sources of multi-source data}
}
\value{
S: Results of variable selection;
FDRPower: FDR and Power.
}
\description{
The values of FDR and Power under Aggregating multiple for SKMCD (CV).
}
