% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{fit_marginals}
\alias{fit_marginals}
\title{Fit the marginal distributions for each row of a count matrix (Reference package 'scDesign2').}
\usage{
fit_marginals(
  x,
  marginal = c("auto_choose", "zinb", "nb", "poisson"),
  pval_cutoff = 0.05,
  epsilon = 1e-05,
  jitter = TRUE,
  DT = TRUE
)
}
\arguments{
\item{x}{A matrix of shape p by n that contains count values.}

\item{marginal}{Specification of the types of marginal distribution.
Default value is 'auto_choose' which chooses between ZINB, NB, ZIP
and Poisson by a likelihood ratio test (lrt) and whether there is
underdispersion.
'zinb' will fit the ZINB model. If there is underdispersion, it
will choose between ZIP and Poisson by a lrt. Otherwise, it will try to
fit the ZINB model. If in this case, there is no zero at all or an error
occurs, it will fit an NB model instead.
'nb' fits the NB model that chooses between NB and Poisson depending
on whether there is underdispersion.
'poisson' simply fits the Poisson model.}

\item{pval_cutoff}{Cutoff of p-value of the lrt that determines whether
there is zero inflation.}

\item{epsilon}{Threshold value for preventing the transformed quantile
to collapse to 0 or 1.}

\item{jitter}{Logical, whether a random projection should be performed in the
distributional transform.}

\item{DT}{Logical, whether distributional transformed should be performed.
If set to FALSE, the returned object u will be NULL.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        a list with the following components:
}\if{html}{\out{</div>}}
}
\description{
Fit the marginal distributions for each row of a count matrix (Reference package 'scDesign2').
}
