% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{inter_cw}
\alias{inter_cw}
\title{The values of FDR and Power under intersection were calculated (CV).}
\usage{
inter_cw(c_w, b_1 = NULL, fdr = 0.2)
}
\arguments{
\item{c_w}{Test statistics;}

\item{b_1}{True significant variable.}

\item{fdr}{Target control level.}
}
\value{
S: Results of variable selection;
FDRPower: FDR and Power.
}
\description{
The values of FDR and Power under intersection were calculated (CV).
}
