% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{konckoffs_simulation}
\alias{konckoffs_simulation}
\title{knockoff generation based on Model-X methods.}
\usage{
konckoffs_simulation(W, class0)
}
\arguments{
\item{W}{Count data.}

\item{class0}{Response variable, binary variable or others.}
}
\value{
result: knockoff generation.
}
\description{
knockoff generation based on Model-X methods.
}
