% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_SK.R
\name{result_cv}
\alias{result_cv}
\title{Title}
\usage{
result_cv(r_ZIPG_SK_cv, B = 1, filter_statistics = NULL, In = F, b_1 = NULL)
}
\arguments{
\item{b_1}{}
}
\value{
SS1: Results based on the statistic of fliter in the method in the CV case
FP: Same as FDRPower;
SS_f1: results based on other fliter's statistics in the CV case.
}
\description{
Title
}
