% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{simulate_count_copula_3}
\alias{simulate_count_copula_3}
\title{knockoff generation based on Gauss-copula and ZIPG methods.}
\usage{
simulate_count_copula_3(copula_result, data_x, M)
}
\arguments{
\item{copula_result}{Parameter estimation results of the 'simulate_data_2';}

\item{data_x}{Data on covariates;}

\item{M}{Count sequencing depth.}
}
\value{
result3: count data.
}
\description{
knockoff generation based on Gauss-copula and ZIPG methods.
}
\examples{
data(data_pg_copula)
data_K_j <- data_pg_copula[[1]][[1]][[1]]
count_K_j <- data_K_j[data_K_j[, 1] == 1, ]
n_x1 <- ifelse(n_x1 == 0, 3, n_x1)
data_x <- as.data.frame(count_K_j[, c(1:n_x1 + 1)])
W <- as.data.frame(count_K_j[, -c(1:(n_x1 + 2))])
M <- count_K_j[, n_x1 + 2]

copula_result <- ZIPG_Estimate_3_1(data_x, W, M)
result3 <- simulate_count_copula_3(copula_result, data_x, M)

}
