% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{simulate_data_2}
\alias{simulate_data_2}
\title{Code to generate a simulated count data.}
\usage{
simulate_data_2(
  n_1 = 100,
  n_p = 100,
  rho_p = 0.6,
  theta,
  lambda,
  prob_0,
  copula = TRUE,
  marginal = "pg"
)
}
\arguments{
\item{n_1}{same as 'simulate_PG_6';}

\item{theta}{The abundance dispersion parameter;}

\item{lambda}{The abundance mean parameter;}

\item{marginal}{}
}
\value{
result3: count data.
}
\description{
Code to generate a simulated count data.
}
