% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{simulate_PG_6}
\alias{simulate_PG_6}
\title{Code for generating data such as one simulated count data and its covariates}
\usage{
simulate_PG_6(
  n_1 = 20,
  n_p = 10,
  n_p_1 = 5,
  diff = 0,
  k = 1,
  prob_up1 = 0.1,
  prob_up2 = 0.1,
  rho_p = 0.5,
  n_x = 3,
  rho_x = 0.1,
  M_max = 100,
  fc = 1,
  marginal = c("nb", "pg"),
  copula = TRUE,
  seed = 111
)
}
\arguments{
\item{n_1}{a scalar, same as 'simulate_datasets_3';}

\item{n_p}{A scalar;}

\item{n_p_1}{A scalar;}

\item{k}{diff relative value, if diff is a vector, k is the corresponding value;}

\item{prob_up1}{The maximum zero proportion for a significant variable}

\item{prob_up2}{The maximum zero proportion of insignificant variables;}

\item{rho_p}{Calculation parameters of covariance matrix based on copula mode;}

\item{n_x}{Number of covariates;}

\item{rho_x}{Parameters of the way in which the covariates were generated.}

\item{M_max}{The mean value of count data;}

\item{fc}{Mean differences of significant variables in different distributions;}

\item{seed}{Value of seed.}
}
\value{
W_simulate: Simulated count data;
marginal: Marginal distribution;
data_x: Data on covariates;
M: Depth of count;
beta: The parameters of the covariates;
prob_0: The zero scale of the simulation.
}
\description{
Code for generating data such as one simulated count data and its covariates
}
