% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{SparseDOSSA_simulation}
\alias{SparseDOSSA_simulation}
\title{knockoff generation based on SparseDOSSA methods.}
\usage{
SparseDOSSA_simulation(W)
}
\arguments{
\item{W}{Count data.}
}
\value{
Stool_simulation_data: knockoff generation.
}
\description{
knockoff generation based on SparseDOSSA methods.
}
