% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/others.R
\name{ZIPG_Estimate_3_1}
\alias{ZIPG_Estimate_3_1}
\title{Parameter estimation based on Gauss-copula and ZIPG methods.}
\usage{
ZIPG_Estimate_3_1(
  data_x,
  W,
  M,
  zp_cutoff = 0.8,
  min_nonzero_num = 1,
  epsilon = 1e-05
)
}
\arguments{
\item{data_x}{Data on covariates;}

\item{W}{count data;}

\item{M}{Depth of count;}

\item{zp_cutoff}{Zero proportion threshold, beyond a certain threshold does not participate in the parameter estimation of the Gaussian copula;}

\item{min_nonzero_num}{Minimum number of nonzero values, avoiding all zeros;}

\item{epsilon}{Threshold of error.}
}
\value{
cov_mat: Covariance of the estimated Gaussian copula model;
marginal_param1: Copula based parameter estimation of count variables not exceeding zp_cutoff threshold;
marginal_param2: Parameter estimation of count variables without copula;
gene_sel1: Copula-based set of variables;
gene_sel2: Variable sets not based on copula;
gene_sel3: A set of variables with all zero counts.
}
\description{
Parameter estimation based on Gauss-copula and ZIPG methods.
}
\examples{
data(data_pg_copula)
data_K_j <- data_pg_copula[[1]][[1]][[1]]
count_K_j <- data_K_j[data_K_j[, 1] == 1, ]
n_x1 <- ifelse(n_x1 == 0, 3, n_x1)
data_x <- as.data.frame(count_K_j[, c(1:n_x1 + 1)])
W <- as.data.frame(count_K_j[, -c(1:(n_x1 + 2))])
M <- count_K_j[, n_x1 + 2]

copula_result <- ZIPG_Estimate_3_1(data_x, W, M)
result3 <- simulate_count_copula_3(copula_result, data_x, M)

}
