% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_SK.R
\name{ZIPG_SK_B}
\alias{ZIPG_SK_B}
\title{An Aggregating version of ZIPG_SK}
\usage{
ZIPG_SK_B(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  B = 1,
  Bstat = 2,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test",
  combine_1 = "simul"
)
}
\arguments{
\item{W}{Same as ZIPG_SK;}

\item{class_K}{Matrix data, representing the input of count data;}

\item{data_x}{Vector data, representing different sources of multi-source data; default is NULL, representing one source or pooling method;}

\item{M}{The sequencing depth of count data, which was the row sum of matrix data;}

\item{y}{Response variable, representing binary, continuous or mixed variables;}

\item{T_var}{True correlation variable, which represents the set of variables that are truly correlated with the response variable, defaults to NULL;}

\item{fdr}{the target FDR level, default 0.2;}

\item{offset}{value between 0 and 1;}

\item{B}{Number of aggregations;}

\item{Bstat}{For aggregation, 1 represents the aggregated test statistic and 2 represents the aggregated fliter statistic;}

\item{test_statistic}{Statistics for a single source dataset, c('DE','GLM','RF')}

\item{filter_statistics}{The statistics of the dataset were calculated from multiple sources. Denote c(cumprod,max, sum) by 1,2,3;}

\item{test1}{Test method}

\item{combine_1}{When combine_1 aggregation mode is 1, 'simul' means that the statistic is calculated based on the proposed method, and 'fisher' means that the statistic is calculated based on the combination of p values.}
}
\value{
S: The variable selection set;
c_w_pis: Same as c_w;
contrast_score: fliter statistics for the non-aggregated case (B=1);
FDRPower: The resulting FDR and Power are calculated after given T_var.
chis: The threshold of the chi-square distribution using fisher's method in the first aggregation case (Bstat = 1, combine_1 == 'fisher');
p_comb:
e_value: Combined E-values for multi-source datasets;
e_w_B: The e value of the multi-source dataset in the aggregated case, B times;
e_value: E-values for single-source datasets.
}
\description{
An Aggregating version of ZIPG_SK
}
