% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_SK.R
\name{ZIPG_SK_cv}
\alias{ZIPG_SK_cv}
\title{ZIPG_SK method based on the cross-validation case}
\usage{
ZIPG_SK_cv(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  cv = 1,
  method = "ZIPG",
  B = 1,
  Bstat = 2,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test",
  combine_1 = "simul",
  In = F
)
}
\arguments{
\item{W}{Same as ZIPG_SK;}

\item{class_K}{Matrix data, representing the input of count data;}

\item{data_x}{Vector data, representing different sources of multi-source data; default is NULL, representing one source or pooling method;}

\item{M}{The sequencing depth of count data, which was the row sum of matrix data;}

\item{y}{Response variable, representing binary, continuous or mixed variables;}

\item{T_var}{True correlation variable, which represents the set of variables that are truly correlated with the response variable, defaults to NULL;}

\item{fdr}{the target FDR level, default 0.2;}

\item{offset}{value between 0 and 1;}

\item{cv}{Fold number of cross validation;}

\item{method}{Synthetic data generation methods, c('ZIPG','ZINB','knockoff');}

\item{B}{Number of aggregations;}

\item{Bstat}{For aggregation, 1 represents the aggregated test statistic and 2 represents the aggregated fliter statistic;}

\item{test_statistic}{Statistics for a single source dataset, c('DE','GLM','RF')}

\item{filter_statistics}{The statistics of the dataset were calculated from multiple sources. Denote c(cumprod,max, sum) by 1,2,3;}

\item{test1}{Test method}

\item{combine_1}{When combine_1 aggregation mode is 1, 'simul' means that the statistic is calculated based on the proposed method, and 'fisher' means that the statistic is calculated based on the combination of p values.}

\item{In}{Boolean value that represents the result of the computation based on the Intersection.}
}
\value{
Same as ZIPG_SK and ZIPG_SK_B;
}
\description{
ZIPG_SK method based on the cross-validation case
}
