% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPG_SK.R
\name{ZIPG_SK_other}
\alias{ZIPG_SK_other}
\title{Extension of ZIPG_SK to consider more synthetic data methods.}
\usage{
ZIPG_SK_other(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  method = "ZIPG",
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test"
)
}
\arguments{
\item{W}{Same as ZIPG_SK;}

\item{method}{Synthetic data generation methods, c('ZIPG','ZINB','knockoff');}

\item{test1}{}
}
\value{
Same as ZIPG_SK;
}
\description{
Extension of ZIPG_SK to consider more synthetic data methods.
}
\examples{

data(data_pg_copula)
i <- 1 # or 2
data_K_j <- data_pg_copula[[1]][[1]][[i]][[1]][[1]]
count_K_j <- data_K_j # [data_K_j[,1]==c,]
n_x1 <- 3
data_x <- as.data.frame(count_K_j[, c(1:n_x1 + 1)])
W <- as.data.frame(count_K_j[, -c(1:(n_x1 + 2))])
M <- count_K_j[, n_x1 + 2]
n_w <- dim(W)[1]
n_data <- 2
y <- rep(rep(c(1, 2), n_data), rep(n_w / 2 / n_data, n_data * 2))
class_K <- rep(c(1:n_data), rep(n_w / n_data, n_data))
n_p_all <- c(50, 100, 400)
n_p <- c(10, 20, 40)
T_var <- 1:n_p[i]
name_data <- names(table(class_K))
fdr <- 0.2

ZIPG_DE_S3 <- ZIPG_SK_other(
  W = W, class_K = class_K, data_x = data_x, M = M, y = y, T_var = T_var,
  fdr = fdr, method = "ZIPG", test_statistic = "DE", filter_statistics = 3
)
}
